osbuild2: move v1 result handling code to separate file
